initial commit for InlineCompletionProvider
