+++ extensions/intellij/gradle.properties
@@ -5,10 +5,11 @@ pluginRepositoryUrl=https://github.com/continuedev/continue
 # SemVer format -> https://semver.org
 pluginVersion=1.0.14
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
-pluginSinceBuild=223
+pluginSinceBuild=232
+pluginUntilBuild=243.*
 # IntelliJ Platform Properties -> https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#configuration-intellij-extension
 platformType=IC
-platformVersion=2022.3.3
+platformVersion=2023.2
 #org.gradle.java.home=/opt/homebrew/opt/openjdk@17
 #platformVersion = LATEST-EAP-SNAPSHOT
 # Plugin Dependencies -> https://plugins.jetbrains.com/docs/intellij/plugin-dependencies.html
